Build script fix for macos

+++ build/cmake/build.sh
@@ -3,7 +3,7 @@
 
 # Build script for building 'MasterSimulator' and all dependend libraries
 
-# Command line options:  
+# Command line options:
 #   [reldeb|release|debug]		build type
 #   [2 [1..n]]					cpu count
 #   [gcc|icc]					compiler
@@ -44,7 +44,7 @@ do
 		echo "  [verbose]                   enable cmake to call verbose makefiles"
 		echo "  [skip-test]                 does not execute test script after successful build"
 
-		exit 
+		exit
 	fi
 
     if [[ $var = *[[:digit:]]* ]];
@@ -52,32 +52,32 @@ do
 		MAKE_CPUCOUNT=$var
 		echo "Using $MAKE_CPUCOUNT CPUs for compilation"
     fi
-    
-    if [[ $var = "omp"  ]]; 
+
+    if [[ $var = "omp"  ]];
     then
 		CMAKE_OPTIONS="$CMAKE_OPTIONS -DUSE_OMP:BOOL=ON"
 		echo "Using Open MP compile flags"
     fi
 
-    if [[ $var = "debug"  ]]; 
+    if [[ $var = "debug"  ]];
     then
 		CMAKE_BUILD_TYPE=" -DCMAKE_BUILD_TYPE:STRING=Debug"
 		echo "Debug build..."
     fi
 
-    if [[ $var = "release"  ]]; 
+    if [[ $var = "release"  ]];
     then
 		CMAKE_BUILD_TYPE=" -DCMAKE_BUILD_TYPE:STRING=Release"
 		echo "Release build..."
     fi
 
-    if [[ $var = "reldeb"  ]]; 
+    if [[ $var = "reldeb"  ]];
     then
 		CMAKE_BUILD_TYPE=" -DCMAKE_BUILD_TYPE:STRING=RelWithDebInfo"
 		echo "RelWithDebInfo build..."
     fi
 
-    if [[ $var = "icc"  && $COMPILER = "" ]]; 
+    if [[ $var = "icc"  && $COMPILER = "" ]];
     then
 		COMPILER="icc"
 		BUILD_DIR_SUFFIX="icc"
@@ -86,7 +86,7 @@ do
 	    CMAKE_COMPILER_OPTIONS="-DCMAKE_C_COMPILER=icc -DCMAKE_CXX_COMPILER=icc"
 	  fi
 
-    if [[ $var = "gcc"  && $COMPILER = "" ]]; 
+    if [[ $var = "gcc"  && $COMPILER = "" ]];
     then
 		COMPILER="gcc"
 		BUILD_DIR_SUFFIX="gcc"
@@ -94,22 +94,22 @@ do
 		CMAKE_COMPILER_OPTIONS=""
 	  fi
 
-    if [[ $var = "verbose"  ]]; 
+    if [[ $var = "verbose"  ]];
   	then
 		CMAKE_OPTIONS="$CMAKE_OPTIONS -DCMAKE_VERBOSE_MAKEFILE:BOOL=ON"
 	  fi
-	  
-    if [[ $var = "skip-test"  ]]; 
+
+    if [[ $var = "skip-test"  ]];
   	then
 		SKIP_TESTS="true"
 	  fi
 
-    if [[ $var = "lapack"  ]]; 
+    if [[ $var = "lapack"  ]];
     then
 		CMAKE_OPTIONS="$CMAKE_OPTIONS -DLAPACK_ENABLE:BOOL=ON"
 		echo "Building with lapack support for CVODE"
     fi
-	  
+
 done
 
 
@@ -125,7 +125,7 @@ do
 		echo "Gprof build, forcing GCC build..."
     fi
 
-    if [[ $var = "threadChecker"  ]]; 
+    if [[ $var = "threadChecker"  ]];
     then
 		COMPILER="icc"
 		BUILD_DIR_SUFFIX="icc"
@@ -143,7 +143,7 @@ if [ ! -d $BUILDDIR ]; then
     mkdir -p $BUILDDIR
 fi
 
-cd $BUILDDIR && cmake $CMAKE_OPTIONS $CMAKE_BUILD_TYPE $CMAKE_COMPILER_OPTIONS $CMAKELISTSDIR && make -j$MAKE_CPUCOUNT && 
+cd $BUILDDIR && cmake $CMAKE_OPTIONS $CMAKE_BUILD_TYPE $CMAKE_COMPILER_OPTIONS $CMAKELISTSDIR && make -j$MAKE_CPUCOUNT &&
 cd $CMAKELISTSDIR &&
 mkdir -p ../../bin/release &&
 if [ -e $BUILDDIR/MasterSimulator/MasterSimulator ]; then
@@ -156,21 +156,25 @@ if [ -e $BUILDDIR/MasterSimulatorUI/MasterSimulatorUI ]; then
 fi &&
 if [ -e $BUILDDIR/MasterSimulatorUI/MasterSimulatorUI.app ]; then
   if [ -e ../../bin/release/MasterSimulatorUI.app ]; then
-    rm -rf ../../bin/release/MasterSimulatorUI.app 
+    rm -rf ../../bin/release/MasterSimulatorUI.app
   fi &&
   echo "*** Copying MasterSimulatorUI.app to bin/release ***" &&
-  cp -r $BUILDDIR/MasterSimulatorUI/MasterSimulatorUI.app ../../bin/release/MasterSimulatorUI.app 
+  cp -r $BUILDDIR/MasterSimulatorUI/MasterSimulatorUI.app ../../bin/release/MasterSimulatorUI.app
 fi &&
 
 echo "*** TestFMUs to bin/release ***" &&
-cp $BUILDDIR/Math003Part1/libMath003Part1.so ../../bin/release/libMath003Part1.so &&
-cp $BUILDDIR/Math003Part2/libMath003Part2.so ../../bin/release/libMath003Part2.so &&
-cp $BUILDDIR/Math003Part3/libMath003Part3.so ../../bin/release/libMath003Part3.so &&
+if [ -e $BUILDDIR/MasterSimulatorUI/MasterSimulatorUI.app ]; then
+	cp $BUILDDIR/Math003Part1/libMath003Part1.dylib ../../bin/release/libMath003Part1.dylib &&
+  cp $BUILDDIR/Math003Part2/libMath003Part2.dylib ../../bin/release/libMath003Part2.dylib &&
+  cp $BUILDDIR/Math003Part3/libMath003Part3.dylib ../../bin/release/libMath003Part3.dylib
+else
+  cp $BUILDDIR/Math003Part1/libMath003Part1.so ../../bin/release/libMath003Part1.so &&
+  cp $BUILDDIR/Math003Part2/libMath003Part2.so ../../bin/release/libMath003Part2.so &&
+  cp $BUILDDIR/Math003Part3/libMath003Part3.so ../../bin/release/libMath003Part3.so
+fi &&
 
 echo "*** Build MasterSimulator ***" &&
-if [[ $SKIP_TESTS = "false"  ]]; 
+if [[ $SKIP_TESTS = "false"  ]];
 then
 ./run_tests.sh
 fi
-
-
